fix(billing): Split usage module to avoid server imports in client bundle
Move Amberflo fetching into usage.server.ts and keep pure helpers in
usage-summary.ts so usage routes do not pull evaluate.server into the client build.
