+++ app/modules/billing/usage.server.ts
@@ -1,3 +1,4 @@
+import type { MeterDefinition, MeterSeries, UsageFetchResult } from './usage.types';
 import { client } from '@/modules/control-plane/shared/client.gen';
 import { FeatureFlag } from '@/modules/feature-flags';
 import { isFeatureEnabled } from '@/modules/feature-flags/evaluate.server';
@@ -7,40 +8,7 @@ import { createProjectService } from '@/resources/projects';
 import { env } from '@/utils/env/env.server';
 import { AuthenticationError, AuthorizationError } from '@/utils/errors';
 
-export interface MeterSeries {
-  meterApiName: string;
-  label: string;
-  values: { timestamp: number; value: number }[];
-}
-
-export interface MeterDefinition {
-  meterName: string;
-  displayName: string;
-}
-
-export type UsageFetchStatus =
-  | 'ok'
-  | 'unconfigured'
-  | 'insufficient-permissions'
-  | 'no-billing-account'
-  | 'feature-disabled';
-
-export interface UsageFetchResult {
-  status: UsageFetchStatus;
-  meters: MeterSeries[];
-  days: number;
-  message?: string;
-}
-
-export interface SummarizedMeterUsage {
-  meterApiName: string;
-  label: string;
-  total: number;
-  recentDaily: { date: string; value: number }[];
-}
-
 const DEFAULT_DAYS = 30;
-const ASSISTANT_DAILY_DETAIL_DAYS = 7;
 
 export async function listMeterDefinitions(): Promise<MeterDefinition[]> {
   try {
@@ -234,36 +202,4 @@ export async function fetchOrgUsage(orgId: string, days = DEFAULT_DAYS): Promise
 
   const meters = await fetchUsageForCustomerIds({ customerIds, days });
   return { status: 'ok', meters, days };
-}
-
-export function sumMeterValues(values: { value: number }[]): number {
-  return values.reduce((acc, v) => acc + v.value, 0);
-}
-
-/** Collapse sparse samples into daily totals for assistant context. */
-export function summarizeMetersForAssistant(
-  meters: MeterSeries[],
-  detailDays = ASSISTANT_DAILY_DETAIL_DAYS
-): SummarizedMeterUsage[] {
-  const cutoffMs = Date.now() - detailDays * 24 * 3600 * 1000;
-
-  return meters.map((meter) => {
-    const dailyByDate = new Map<string, number>();
-    for (const { timestamp, value } of meter.values) {
-      if (timestamp < cutoffMs) continue;
-      const date = new Date(timestamp).toISOString().slice(0, 10);
-      dailyByDate.set(date, (dailyByDate.get(date) ?? 0) + value);
-    }
-
-    const recentDaily = Array.from(dailyByDate.entries())
-      .map(([date, value]) => ({ date, value }))
-      .sort((a, b) => a.date.localeCompare(b.date));
-
-    return {
-      meterApiName: meter.meterApiName,
-      label: meter.label,
-      total: sumMeterValues(meter.values),
-      recentDaily,
-    };
-  });
 }
